Remove Filter and replace EndpointCompiled with Kleisli type alias
